+++ gateway/run.py
@@ -1823,144 +1823,13 @@ _TOOL_MEDIA_RE = re.compile(
 )
 
 
-_COMPUTER_USE_CAPTURE_BASENAME_RE = re.compile(
-    r"^computer_use_[0-9a-f]{32}\.(?:png|jpe?g)$",
-    re.IGNORECASE,
+# Shared with cron delivery and gateway background tasks — the repair must
+# run on every surface that feeds a final response into media extraction.
+# Re-exported under the historical private name for tests and callers.
+from gateway.media_repair import (  # noqa: E402
+    repair_explicit_computer_use_media_paths as _repair_explicit_computer_use_media_paths,
+    tool_name_by_call_id as _tool_name_by_call_id,
 )
-_COMPUTER_USE_CAPTURE_SUMMARY_RE = re.compile(
-    r"\(shareable screenshot saved to "
-    r"(?P<path>(?:[A-Za-z]:[/\\]|/|\\\\)[^\r\n]*?"
-    r"computer_use_[0-9a-f]{32}\.(?:png|jpe?g))\)",
-    re.IGNORECASE,
-)
-
-
-def _computer_use_capture_basename(path: Any) -> str:
-    """Return a canonical capture basename for either path separator style."""
-    value = str(path or "").strip().strip("`\"'")
-    basename = re.split(r"[/\\]", value)[-1]
-    if _COMPUTER_USE_CAPTURE_BASENAME_RE.fullmatch(basename):
-        return basename.lower()
-    return ""
-
-
-def _iter_computer_use_capture_paths(content: Any):
-    """Yield persisted screenshot paths from computer_use result content.
-
-    The tool can return JSON, a multimodal content list, or a text fallback.
-    The latter two retain the canonical path in the human-readable summary
-    even though the multimodal envelope's ``meta`` dictionary is not stored in
-    the tool message.
-    """
-    if isinstance(content, str):
-        for match in _COMPUTER_USE_CAPTURE_SUMMARY_RE.finditer(content):
-            yield match.group("path").strip()
-        stripped = content.strip()
-        if stripped.startswith(("{", "[")):
-            try:
-                payload = json.loads(stripped)
-            except Exception:
-                payload = None
-            if isinstance(payload, (dict, list)):
-                yield from _iter_computer_use_capture_paths(payload)
-        return
-
-    if isinstance(content, list):
-        for part in content:
-            yield from _iter_computer_use_capture_paths(part)
-        return
-
-    if not isinstance(content, dict):
-        return
-
-    screenshot_path = content.get("screenshot_path")
-    if isinstance(screenshot_path, str):
-        yield screenshot_path
-    meta = content.get("meta")
-    if isinstance(meta, dict) and isinstance(meta.get("screenshot_path"), str):
-        yield meta["screenshot_path"]
-    for field in ("content", "text", "text_summary", "summary"):
-        nested = content.get(field)
-        if isinstance(nested, (str, dict, list)):
-            yield from _iter_computer_use_capture_paths(nested)
-
-
-def _repair_explicit_computer_use_media_paths(
-    response: str,
-    messages: List[Dict[str, Any]],
-    history_offset: int = 0,
-) -> str:
-    """Recover model-mangled paths for explicitly requested screenshots.
-
-    ``computer_use`` persists a bounded screenshot and gives its absolute path
-    to the model. Some models rewrite a Windows path into a POSIX-looking path
-    before emitting ``MEDIA:``, which makes the gateway reject the nonexistent
-    path. Repair only an already-explicit directive whose unique generated
-    basename matches a canonical screenshot path from this turn. This does not
-    auto-attach ordinary computer-use captures, and normal media path
-    validation still runs after the repair.
-    """
-    if "MEDIA:" not in response:
-        return response
-
-    if history_offset and len(messages) >= history_offset:
-        turn_messages = messages[history_offset:]
-    elif history_offset:
-        # Compression can invalidate the original slice boundary. Recover the
-        # current turn from its last user message; fail closed if none remains.
-        last_user = next(
-            (
-                index
-                for index in range(len(messages) - 1, -1, -1)
-                if messages[index].get("role") == "user"
-            ),
-            None,
-        )
-        turn_messages = messages[last_user:] if last_user is not None else []
-    else:
-        turn_messages = messages
-
-    tool_name_by_call_id: Dict[str, str] = {}
-    for msg in turn_messages:
-        if msg.get("role") != "assistant":
-            continue
-        for call in msg.get("tool_calls") or []:
-            call_id = call.get("id") or call.get("call_id")
-            fn = call.get("function") or {}
-            name = str(fn.get("name") or call.get("name") or "")
-            if call_id and name:
-                tool_name_by_call_id[str(call_id)] = name
-
-    canonical_by_basename: Dict[str, str] = {}
-    for msg in turn_messages:
-        if msg.get("role") not in {"tool", "function"}:
-            continue
-        call_id = str(msg.get("tool_call_id") or msg.get("call_id") or "")
-        tool_name = str(
-            msg.get("name")
-            or msg.get("tool_name")
-            or tool_name_by_call_id.get(call_id)
-            or ""
-        )
-        if tool_name != "computer_use":
-            continue
-        for path in _iter_computer_use_capture_paths(msg.get("content")):
-            basename = _computer_use_capture_basename(path)
-            if basename and re.match(r"^(?:[A-Za-z]:[/\\]|/|\\\\)", path):
-                canonical_by_basename[basename] = path
-
-    if not canonical_by_basename:
-        return response
-
-    media_files, _ = BasePlatformAdapter.extract_media(response)
-    repaired = response
-    for emitted_path, _is_voice in media_files:
-        canonical = canonical_by_basename.get(
-            _computer_use_capture_basename(emitted_path)
-        )
-        if canonical and emitted_path != canonical:
-            repaired = repaired.replace(emitted_path, canonical)
-    return repaired
 
 
 def _collect_auto_append_media_tags(
@@ -1994,16 +1863,7 @@ def _collect_auto_append_media_tags(
     else:
         new_messages = messages
 
-    tool_name_by_call_id: Dict[str, str] = {}
-    for msg in new_messages:
-        if msg.get("role") != "assistant":
-            continue
-        for call in msg.get("tool_calls") or []:
-            call_id = call.get("id") or call.get("call_id")
-            fn = call.get("function") or {}
-            name = str(fn.get("name") or call.get("name") or "")
-            if call_id and name:
-                tool_name_by_call_id[str(call_id)] = name
+    tool_name_by_call_id = _tool_name_by_call_id(new_messages)
 
     media_tags: List[str] = []
     has_voice_directive = False
@@ -2058,7 +1918,7 @@ def _collect_history_media_paths(agent_history: List[Dict[str, Any]]) -> set:
     shape caused repeated delivery when the model echoed a previous MEDIA tag.
     """
     paths: set = set()
-    tool_name_by_call_id: Dict[str, str] = {}
+    tool_name_by_call_id = _tool_name_by_call_id(agent_history)
 
     def _add_text_media_paths(content: str) -> None:
         for match in _TOOL_MEDIA_RE.finditer(content):
@@ -2072,14 +1932,6 @@ def _collect_history_media_paths(agent_history: List[Dict[str, Any]]) -> set:
         media_files, _ = BasePlatformAdapter.extract_media(content)
         paths.update(path for path, _is_voice in media_files)
 
-    for msg in agent_history:
-        if msg.get("role") == "assistant":
-            for call in msg.get("tool_calls") or []:
-                cid = call.get("id") or call.get("call_id")
-                fn = call.get("function") or {}
-                name = str(fn.get("name") or call.get("name") or "")
-                if cid and name:
-                    tool_name_by_call_id[str(cid)] = name
     for msg in agent_history:
         role = msg.get("role")
         if role == "assistant":
@@ -22933,6 +22785,15 @@ class GatewayRunner(GatewayAuthorizationMixin, GatewayKanbanWatchersMixin, Gatew
             response = result.get("final_response", "") if result else ""
             if not response and result and result.get("error"):
                 response = f"Error: {result['error']}"
+
+            # Background tasks start a fresh conversation (no prior history),
+            # so history_offset=0: every message in the run belongs to this
+            # turn. Mirrors the repair on the main turn path.
+            if response and result:
+                response = _repair_explicit_computer_use_media_paths(
+                    response,
+                    result.get("messages", []),
+                )
 
             # Extract media files from the response
             if response:
