+++ tests/gateway/test_media_extraction.py
@@ -214,6 +214,73 @@ class TestMediaExtraction:
 
         assert repaired == response
 
+    def test_malformed_json_result_fails_closed(self):
+        """Truncated JSON must not repair to a doubled-backslash artifact.
+
+        JSON escaping doubles backslashes; regex-scanning the raw string
+        would yield ``C:\\\\Users\\\\...`` — a path that exists nowhere. When
+        json.loads fails, the helper must yield nothing rather than rewrite
+        the response to a corrupted path.
+        """
+        import json as _json
+
+        from gateway.run import _repair_explicit_computer_use_media_paths
+
+        capture_name = "computer_use_dddddddddddddddddddddddddddddddd.png"
+        canonical = rf"C:\Users\Alice\AppData\Local\hermes\cache\images\{capture_name}"
+        payload = _json.dumps(
+            {
+                "summary": f"capture\n  (shareable screenshot saved to {canonical})",
+                "screenshot_path": canonical,
+            }
+        )
+        truncated = payload[:-5]  # starts with '{' but no longer parses
+        response = f"MEDIA:/Users/Alice/AppData/Local/hermes/cache/images/{capture_name}"
+        messages = [
+            {"role": "tool", "name": "computer_use", "content": truncated},
+        ]
+
+        assert (
+            _repair_explicit_computer_use_media_paths(response, messages)
+            == response
+        )
+
+    def test_compression_fallback_slices_from_last_user_message(self):
+        """When compression shrinks messages below history_offset, the repair
+        recovers the current turn from the last user message — and fails
+        closed (no rewrite) when no user message remains."""
+        import json as _json
+
+        from gateway.run import _repair_explicit_computer_use_media_paths
+
+        capture_name = "computer_use_eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee.png"
+        canonical = rf"C:\cache\images\{capture_name}"
+        response = f"MEDIA:/cache/images/{capture_name}"
+        current_turn = [
+            {"role": "user", "content": "Send the screenshot."},
+            {
+                "role": "tool",
+                "name": "computer_use",
+                "content": _json.dumps({"screenshot_path": canonical}),
+            },
+        ]
+
+        # history_offset larger than the (compressed) message list forces the
+        # fallback branch; the last-user slice still finds this turn's result.
+        repaired = _repair_explicit_computer_use_media_paths(
+            response, current_turn, history_offset=10
+        )
+        assert repaired == f"MEDIA:{canonical}"
+
+        # No user message at all -> fail closed, nothing rewritten.
+        no_user = [current_turn[1]]
+        assert (
+            _repair_explicit_computer_use_media_paths(
+                response, no_user, history_offset=10
+            )
+            == response
+        )
+
     def test_gateway_auto_append_ignores_media_examples_in_skill_docs(self):
         """Skill/documentation examples must not be appended as real attachments."""
         from gateway.run import _collect_auto_append_media_tags
